+++ src/HotChocolate/Core/src/Abstractions/SetState.cs
@@ -1,6 +1,6 @@
 namespace HotChocolate
 {
-    public delegate void SetState<T>(T value);
+    public delegate void SetState<in T>(T value);
 
     public delegate void SetState(object value);
 }
